Fix compile-time warnings.
* find/defs.h (struct predicate): Add a missing const modifier.
* find/find.c (wd_sanity_check): Suppress a warning in #else branch.
(process_dir): Remove an unused variables and statements.
* find/pred.c (pred_context): Use const modifier in the prototype.
* lib/buildcmd.{c,h} (bc_args_exceed_testing_limit): Remove a const
modifier causing unnecessary warnings.
* xargs/xargs.c (main): Add explicit type-casts.

+++ xargs/xargs.c
@@ -555,7 +555,7 @@ main (int argc, char **argv)
 		error (0, 0,
 		       _("warning: value %ld for -s option is too large, "
 			 "using %ld instead"),
-		       arg_size, bc_ctl.posix_arg_size_max);
+		       (long) arg_size, (long) bc_ctl.posix_arg_size_max);
 		arg_size = bc_ctl.posix_arg_size_max;
 	      }
 	    bc_ctl.arg_max = arg_size;
